+++ pyslvs_ui/synthesis/collections/structure_widget.py
@@ -294,7 +294,7 @@ class StructureWidget(QWidget, Ui_Form):
         file_name = self.output_to("atlas edges expression", ["Text file (*.txt)"])
         if not file_name:
             return
-        with open(file_name, 'w', encoding='utf-8') as f:
+        with open(file_name, 'w+', encoding='utf-8') as f:
             f.write('\n'.join(str(g.edges) for g in self.collections))
         self.save_reply_box("edges expression", file_name)